fix: fixed hitting cnonection limit when running tests
TL;DR: There's a maximum number of concurrent connections on a single port. I'm not sure exactly what the limit is, but I was hitting it with my tests. So I just added a check, and if any of the requests fail to establish the TCP connection that the HTTP request will go over, it waits a tenth of a second (long enough for other tests to run) and continues.